refactor: group the terminal output under src/ui/
render, tui and banner are the same layer — how a command's result reaches the terminal —
and they already carried their own submodule directories. Nesting them under src/ui/
leaves the top level to the pipeline and lets `curate`'s picker reach its engine with
`use super::…`.
