+++ src/ui/banner/emoji.rs
