+++ src/ui/banner/sine.rs
